for post word data

+++ webapp/service/model/model_content.go
@@ -11,7 +11,7 @@ const (
 )
 
 type ContentWord struct {
-	ID          int8        `gorm:"column:id"           json:"id"`
+	ID          int         `gorm:"column:id"           json:"id"`
 	Wort        string      `gorm:"column:wort"         json:"Wort"      sql:"not null;unique"  binding:"required"`
 	WortSex     string      `gorm:"column:wort_sex"     json:"WortSex"   sql:"not_null"   binding:"required"`
 	Plural      string      `gorm:"column:plural"       json:"Plural2"    sql:"not null"   binding:"required"`
